Check that a Q-matrix follows the expected structure

The Q-matrix should be 1 row per item and 1 column per attribute, with an optional additional column to store item identifiers. Each value of the Q-matrix should be a 0 or 1 to indicate measurement of the attribute by the given item.

Usage

check_qmatrix(
  x,
  identifier = NULL,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

Arguments

x

The provided Q-matrix to check.

identifier

The provided item identifier, as a character string. If no item identifier is present, the value should be NULL (the default).

arg

The name of the argument.

call

The call stack.

Value

A tibble with one row per item.

Examples

example_qmatrix <- tibble::tibble(item = paste0("item_", 1:5),
                                  att_1 = c(0, 0, 1, 1, 1),
                                  att_2 = c(1, 1, 1, 0, 0))
check_qmatrix(example_qmatrix, identifier = "item")
#> # A tibble: 5 × 3
#>   item   att_1 att_2
#>   <chr>  <int> <int>
#> 1 item_1     0     1
#> 2 item_2     0     1
#> 3 item_3     1     1
#> 4 item_4     1     0
#> 5 item_5     1     0